+++ src/server/initials.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import { Server } from "http";
 import Default_Config from './config';
 import Default_Middlewares from './middlewares';
-import { Config, Db, KeyValString, Middleware } from "./model";
+import { Config, Db, Injector, KeyValString, Middleware } from "./model";
 
 export class Initials {
 
@@ -13,7 +13,7 @@ export class Initials {
 
   db = {} as Db;
   middleware = { ...Default_Middlewares } as Middleware;
-  injectors = {} as Db;
+  injectors = {} as { [key: string]: Injector };
   store = {} as Object;
   config = { ...Default_Config } as Config
   rewriters = {} as KeyValString;